backend/requirements.txt
Django==5.1.1
djangorestframework==3.15.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
celery==5.4.0
redis==5.0.4
requests==2.32.3
openai==1.52.0
django-cors-headers==4.4.0
django-environ==0.11.2
python-dotenv==1.0.1
Pillow>=10.0.0

